Move flavor_name_describe.py into flavor_name_check.py
