+++ Tests/iaas/flavor-naming/flavor-name-check.py
@@ -11,8 +11,7 @@ License: CC-BY-SA 4.0
 
 import sys
 
-from flavor_name_check import CompatLayer, inputflavor
-from flavor_name_describe import prettyname
+from flavor_name_check import CompatLayer, inputflavor, prettyname
 
 
 def usage():
